fixup! Move RaftException to org.apache.ratis.protocol.exceptions
