+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/AlreadyClosedException.java
@@ -17,8 +17,6 @@
  */
 package org.apache.ratis.protocol.exceptions;
 
-import org.apache.ratis.protocol.RaftException;
-
 /**
  * The corresponding object is already closed.
  */
