+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/NotLeaderException.java
@@ -17,7 +17,6 @@
  */
 package org.apache.ratis.protocol.exceptions;
 
-import org.apache.ratis.protocol.RaftException;
 import org.apache.ratis.protocol.RaftGroupMemberId;
 import org.apache.ratis.protocol.RaftPeer;
 import org.apache.ratis.util.Preconditions;
