+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/RaftException.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ratis.protocol;
+package org.apache.ratis.protocol.exceptions;
 
 import java.io.IOException;
