+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/RaftRetryFailureException.java
@@ -18,7 +18,6 @@
 package org.apache.ratis.protocol.exceptions;
 
 import org.apache.ratis.protocol.RaftClientRequest;
-import org.apache.ratis.protocol.RaftException;
 import org.apache.ratis.retry.RetryPolicy;
 
 /**
