+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/StreamException.java
@@ -17,8 +17,6 @@
  */
 package org.apache.ratis.protocol.exceptions;
 
-import org.apache.ratis.protocol.RaftException;
-
 /**
  * A stream is not found in the server.
  */
