+++ ratis-server/src/main/java/org/apache/ratis/server/raftlog/RaftLogIOException.java
@@ -17,7 +17,7 @@
  */
 package org.apache.ratis.server.raftlog;
 
-import org.apache.ratis.protocol.RaftException;
+import org.apache.ratis.protocol.exceptions.RaftException;
 
 /**
  * Exception while reading/writing RaftLog
